--- a/commons/src/main/java/org/dhis2/commons/filters/data/FilterBindings.kt
+++ b/commons/src/main/java/org/dhis2/commons/filters/data/FilterBindings.kt
@@ -115,13 +115,6 @@ fun setSortingIcon(
                             R.drawable.ic_sort_deactivated,
                         ),
                     )
-                else ->
-                    sortingIcon.setImageDrawable(
-                        AppCompatResources.getDrawable(
-                            sortingIcon.context,
-                            R.drawable.ic_sort_deactivated,
-                        ),
-                    )
             }
         }
     }