--- a/commons/src/main/java/org/dhis2/commons/resources/ColorUtils.kt
+++ b/commons/src/main/java/org/dhis2/commons/resources/ColorUtils.kt
@@ -201,7 +201,6 @@ fun Context.getPrimaryColor(colorType: ColorType): Int {
             ColorType.PRIMARY_DARK -> R.attr.colorPrimaryDark
             ColorType.PRIMARY_LIGHT -> R.attr.colorPrimaryLight
             ColorType.PRIMARY -> R.attr.colorPrimary
-            else -> R.attr.colorPrimary
         }
     val typedValue = TypedValue()
     val a = obtainStyledAttributes(typedValue.data, intArrayOf(id))