--- a/gradle.properties
+++ b/gradle.properties
@@ -30,8 +30,15 @@ org.gradle.parallel=true
 # Enable configure on demand.
 org.gradle.configureondemand=true
 
+# Enable Gradle build cache
+org.gradle.caching=true
+
 android.suppressUnsupportedCompileSdk=32
 android.nonTransitiveRClass=false
 android.nonFinalResIds=false
 
 android.enableR8.fullMode=false
+
+# Kotlin compiler flags
+kotlin.daemon.jvmargs=-Xmx2048M
+kotlin.incremental=true